criado contasService - adicionado entidades jpa em contasService - otimizado caching no build do docker - arrumado erros no build.gradle.kts do orchestratorService - criado separação de schemas no banco de dados - arrumado localização errada de RabbitmqConsts.java em /shared

+++ backend/shared/lib/src/main/java/br/ufpr/dac/shared/dto/UsersDto.java
@@ -29,7 +29,7 @@ public class UsersDto {
     private String cpf;
     private Integer estado;
     private String telefone;
-    private Integer salario;
+    private Double salario;
     private List<Endereco> enderecos;
   }
